age-prediction/requirements.txt
pillow
torch_mtcnn
